api-crate/requirements.txt
flask
Flask-RESTful==0.3.8
flask-cors
PyJWT==1.7.1
crash
crate
pyopenssl
pymongo [srv]